Refactor dashboard components and update imports
- Updated import paths for Overview, RecentPlants, QuickActions, and StrainDistribution components to reflect the new directory structure.
- Removed obsolete components from the dashboard, streamlining the codebase.
- Enhanced formatting for improved readability in the DashboardPage component.

These changes improve the organization and maintainability of the dashboard, ensuring a more efficient development process.
