+++ src/components/dashboard/recent-plants.tsx
