+++ src/components/dashboard/strain-distribution.tsx
